6795913: A few remaining wrong casts need to be fixed for building hotspot successfully on Mac OS
Use NULL_WORD in the places where intptr_t is expected due to incompatible types between intptr_t & int32_t

+++ hotspot/src/cpu/x86/vm/templateInterpreter_x86_32.cpp
@@ -382,7 +382,7 @@ void InterpreterGenerator::generate_counter_overflow(Label* do_continue) {
   // indicating if the counter overflow occurs at a backwards branch (non-NULL bcp).
   // The call returns the address of the verified entry point for the method or NULL
   // if the compilation did not complete (either went background or bailed out).
-  __ movptr(rax, (int32_t)false);
+  __ movptr(rax, (intptr_t)false);
   __ call_VM(noreg, CAST_FROM_FN_PTR(address, InterpreterRuntime::frequency_counter_overflow), rax);
 
   __ movptr(rbx, Address(rbp, method_offset));   // restore methodOop
